requirements.txt
Flask>=2.0.0
mysql-connector-python>=8.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
